+++ src/xior.ts
@@ -23,7 +23,7 @@ export class xior {
   static create(options?: XiorRequestConfig): XiorInstance {
     return new xior(options);
   }
-  static VERSION = '0.2.1';
+  static VERSION = '0.2.2';
 
   config?: XiorRequestConfig;
   defaults: XiorInterceptorRequestConfig;